requirements.txt
torch~=2.5.1
transformers
Pillow~=11.1.0
opencv-python~=4.10.0.84
tesseract
flask~=3.1.0
fastapi
uvicorn
pytesseract~=0.3.13
PaddleOCR
camelot-py
pdf2image
numpy~=1.26.4
werkzeug~=3.1.3
itsdangerous~=2.2.0
pydantic~=2.10.4
requests~=2.32.3
filelock~=3.16.1
click~=8.1.8
openpyxl~=3.1.5
pandas~=2.2.3
python-dateutil~=2.9.0.post0
chardet~=5.2.0
scikit-learn~=1.6.0
tqdm~=4.64.0
pypdf~=3.17.4
six~=1.17.0
pyclipper~=1.3.0.post6
shapely~=2.0.6
networkx~=3.4.2
pdf2image~=1.17.0